refactor(blorktools): bounce helper
